@@ -1,118 +1,119 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-	  <Authors>Mohamed Marzooq</Authors>
-	  <TargetFramework>net6.0</TargetFramework>
-	  <ImplicitUsings>enable</ImplicitUsings>
-	  <Nullable>enable</Nullable>
-	  <IsPackable>false</IsPackable>
-	  <IsTestProject>true</IsTestProject>
-	  <LangVersion>latest</LangVersion>
-	  <Configurations>Debug;Release;Dev;Test;Uat;Staging</Configurations>
-  </PropertyGroup>
+	<PropertyGroup>
+		<Authors>Mohamed Marzooq</Authors>
+		<TargetFramework>net9.0</TargetFramework>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Nullable>enable</Nullable>
+		<IsPackable>false</IsPackable>
+		<IsTestProject>true</IsTestProject>
+		<LangVersion>latest</LangVersion>
+		<Configurations>Debug;Release;Dev;Test;Uat;Staging</Configurations>
+	</PropertyGroup>
 
-  <ItemGroup>
-	  <PackageReference Include="BoDi" Version="1.5.0" />
-	  <PackageReference Include="coverlet.collector" Version="6.0.2">
-		<PrivateAssets>all</PrivateAssets>
-		<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-	  </PackageReference>
-	  <PackageReference Include="DotNetSeleniumExtras.WaitHelpers" Version="3.11.0" />
-	  <PackageReference Include="FluentAssertions" Version="6.12.0" />
-	  <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="8.0.0" />
-	  <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.10.0" />
-	  <PackageReference Include="Microsoft.VisualStudio.SlowCheetah" Version="4.0.50">
-		<PrivateAssets>all</PrivateAssets>
-		<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-	  </PackageReference>
-	  <PackageReference Include="nunit" Version="4.1.0" />
-	  <PackageReference Include="NUnit3TestAdapter" Version="4.5.0" />
-	  <PackageReference Include="Roslynator.Analyzers" Version="4.12.4">
-		<PrivateAssets>all</PrivateAssets>
-		<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-	  </PackageReference>
-	  <PackageReference Include="Selenium.SelfHealing.v4" Version="0.99.24-beta.1" />
-	  <PackageReference Include="Selenium.Support" Version="4.23.0">
-		<TreatAsUsed>true</TreatAsUsed>
-	  </PackageReference>
-	  <PackageReference Include="Selenium.WebDriver" Version="4.23.0">
-		<TreatAsUsed>true</TreatAsUsed>
-	  </PackageReference>
-	  <PackageReference Include="Serilog.Sinks.File" Version="6.0.0" />
-	  <PackageReference Include="SpecFlow.NUnit" Version="3.9.74" />
-  </ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="coverlet.collector" Version="6.0.2">
+			<PrivateAssets>all</PrivateAssets>
+			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+		</PackageReference>
+		<PackageReference Include="DotNetSeleniumExtras.WaitHelpers" Version="3.11.0" />
+		<PackageReference Include="FluentAssertions" Version="6.12.0" />
+		<PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="8.0.0" />
+		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.10.0" />
+		<PackageReference Include="Microsoft.VisualStudio.SlowCheetah" Version="4.0.50">
+			<PrivateAssets>all</PrivateAssets>
+			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+		</PackageReference>
+		<PackageReference Include="nunit" Version="4.1.0" />
+		<PackageReference Include="NUnit3TestAdapter" Version="4.5.0" />
+		<PackageReference Include="Reqnroll" Version="2.4.1" />
+		<PackageReference Include="Reqnroll.NUnit" Version="2.4.1" />
+		<PackageReference Include="Reqnroll.SpecFlowCompatibility" Version="2.4.1" />
+		<PackageReference Include="Roslynator.Analyzers" Version="4.12.4">
+			<PrivateAssets>all</PrivateAssets>
+			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+		</PackageReference>
+		<PackageReference Include="Selenium.SelfHealing.v4" Version="0.99.24-beta.1" />
+		<PackageReference Include="Selenium.Support" Version="4.23.0">
+			<TreatAsUsed>true</TreatAsUsed>
+		</PackageReference>
+		<PackageReference Include="Selenium.WebDriver" Version="4.23.0">
+			<TreatAsUsed>true</TreatAsUsed>
+		</PackageReference>
+		<PackageReference Include="Serilog.Sinks.File" Version="6.0.0" />
+	</ItemGroup>
 
-  <ItemGroup>
-	  <Using Include="BoDi" />
-	  <Using Include="FluentAssertions" />
-	  <Using Include="NUnit.Framework" />
-	  <Using Include="OpenQA.Selenium" />
-	  <Using Include="SeleniumExtras.WaitHelpers" />
-	  <Using Include="Serilog" />
-	  <Using Include="System.Diagnostics" />
-	  <Using Include="TechTalk.SpecFlow" />
-  </ItemGroup>
+	<ItemGroup>
+		<Using Include="FluentAssertions" />
+		<Using Include="NUnit.Framework" />
+		<Using Include="OpenQA.Selenium" />
+		<Using Include="SeleniumExtras.WaitHelpers" />
+		<Using Include="Serilog" />
+		<Using Include="System.Diagnostics" />
+		<Using Include="Reqnroll" />
+		<Using Include="Reqnroll.BoDi" />
+	</ItemGroup>
 
-  <ItemGroup>
-	<None Update="Tests\Configuration\Config.Debug.json">
-	  <IsTransformFile>true</IsTransformFile>
-	  <DependentUpon>Config.json</DependentUpon>
-	</None>
-	<None Update="Tests\Configuration\Config.Dev.json">
-	  <IsTransformFile>true</IsTransformFile>
-	  <DependentUpon>Config.json</DependentUpon>
-	</None>
-	<None Update="Tests\Configuration\Config.json">
-	  <TransformOnBuild>true</TransformOnBuild>
-	  <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-	</None>
-	<None Update="Tests\Configuration\Config.Release.json">
-	  <IsTransformFile>true</IsTransformFile>
-	  <DependentUpon>Config.json</DependentUpon>
-	</None>
-	<None Update="Tests\Configuration\Config.Staging.json">
-	  <IsTransformFile>true</IsTransformFile>
-	  <DependentUpon>Config.json</DependentUpon>
-	</None>
-	<None Update="Tests\Configuration\Config.Test.json">
-	  <IsTransformFile>true</IsTransformFile>
-	  <DependentUpon>Config.json</DependentUpon>
-	</None>
-	<None Update="Tests\Configuration\Config.Uat.json">
-	  <IsTransformFile>true</IsTransformFile>
-	  <DependentUpon>Config.json</DependentUpon>
-	</None>
-	<None Update="Utilities\Executors\KillWebDrivers.bat">
-	  <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-	</None>
-	<None Update="Utilities\WebDrivers\BrowserOptions\BrowserSettings.Debug.json">
-	  <IsTransformFile>true</IsTransformFile>
-	  <DependentUpon>BrowserSettings.json</DependentUpon>
-	</None>
-	<None Update="Utilities\WebDrivers\BrowserOptions\BrowserSettings.Dev.json">
-	  <IsTransformFile>true</IsTransformFile>
-	  <DependentUpon>BrowserSettings.json</DependentUpon>
-	</None>
-	<None Update="Utilities\WebDrivers\BrowserOptions\BrowserSettings.json">
-	  <TransformOnBuild>true</TransformOnBuild>
-	  <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-	</None>
-	<None Update="Utilities\WebDrivers\BrowserOptions\BrowserSettings.Release.json">
-	  <IsTransformFile>true</IsTransformFile>
-	  <DependentUpon>BrowserSettings.json</DependentUpon>
-	</None>
-	<None Update="Utilities\WebDrivers\BrowserOptions\BrowserSettings.Staging.json">
-	  <IsTransformFile>true</IsTransformFile>
-	  <DependentUpon>BrowserSettings.json</DependentUpon>
-	</None>
-	<None Update="Utilities\WebDrivers\BrowserOptions\BrowserSettings.Test.json">
-	  <IsTransformFile>true</IsTransformFile>
-	  <DependentUpon>BrowserSettings.json</DependentUpon>
-	</None>
-	<None Update="Utilities\WebDrivers\BrowserOptions\BrowserSettings.Uat.json">
-	  <IsTransformFile>true</IsTransformFile>
-	  <DependentUpon>BrowserSettings.json</DependentUpon>
-	</None>
-  </ItemGroup>
+	<ItemGroup>
+		<None Update="Tests\Configuration\Config.Debug.json">
+			<IsTransformFile>true</IsTransformFile>
+			<DependentUpon>Config.json</DependentUpon>
+		</None>
+		<None Update="Tests\Configuration\Config.Dev.json">
+			<IsTransformFile>true</IsTransformFile>
+			<DependentUpon>Config.json</DependentUpon>
+		</None>
+		<None Update="Tests\Configuration\Config.json">
+			<TransformOnBuild>true</TransformOnBuild>
+			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
+		</None>
+		<None Update="Tests\Configuration\Config.Release.json">
+			<IsTransformFile>true</IsTransformFile>
+			<DependentUpon>Config.json</DependentUpon>
+		</None>
+		<None Update="Tests\Configuration\Config.Staging.json">
+			<IsTransformFile>true</IsTransformFile>
+			<DependentUpon>Config.json</DependentUpon>
+		</None>
+		<None Update="Tests\Configuration\Config.Test.json">
+			<IsTransformFile>true</IsTransformFile>
+			<DependentUpon>Config.json</DependentUpon>
+		</None>
+		<None Update="Tests\Configuration\Config.Uat.json">
+			<IsTransformFile>true</IsTransformFile>
+			<DependentUpon>Config.json</DependentUpon>
+		</None>
+		<None Update="Utilities\Executors\KillWebDrivers.bat">
+			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
+		</None>
+		<None Update="Utilities\WebDrivers\BrowserOptions\BrowserSettings.Debug.json">
+			<IsTransformFile>true</IsTransformFile>
+			<DependentUpon>BrowserSettings.json</DependentUpon>
+		</None>
+		<None Update="Utilities\WebDrivers\BrowserOptions\BrowserSettings.Dev.json">
+			<IsTransformFile>true</IsTransformFile>
+			<DependentUpon>BrowserSettings.json</DependentUpon>
+		</None>
+		<None Update="Utilities\WebDrivers\BrowserOptions\BrowserSettings.json">
+			<TransformOnBuild>true</TransformOnBuild>
+			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
+		</None>
+		<None Update="Utilities\WebDrivers\BrowserOptions\BrowserSettings.Release.json">
+			<IsTransformFile>true</IsTransformFile>
+			<DependentUpon>BrowserSettings.json</DependentUpon>
+		</None>
+		<None Update="Utilities\WebDrivers\BrowserOptions\BrowserSettings.Staging.json">
+			<IsTransformFile>true</IsTransformFile>
+			<DependentUpon>BrowserSettings.json</DependentUpon>
+		</None>
+		<None Update="Utilities\WebDrivers\BrowserOptions\BrowserSettings.Test.json">
+			<IsTransformFile>true</IsTransformFile>
+			<DependentUpon>BrowserSettings.json</DependentUpon>
+		</None>
+		<None Update="Utilities\WebDrivers\BrowserOptions\BrowserSettings.Uat.json">
+			<IsTransformFile>true</IsTransformFile>
+			<DependentUpon>BrowserSettings.json</DependentUpon>
+		</None>
+	</ItemGroup>
 
 </Project>
