@@ -1,7 +1,6 @@
 ﻿[assembly: LevelOfParallelism(4)]
 [assembly: Parallelizable(ParallelScope.Fixtures)]
 
-namespace SeleniumSharpTemplate.Utilities.Attributes
-{
-    internal interface IParallelAttributes;
-}
+namespace SeleniumSharpTemplate.Utilities.Attributes;
+
+internal interface IParallelAttributes;
