@@ -1,14 +1,13 @@
 ﻿using Microsoft.Extensions.Configuration;
 
-namespace SeleniumSharpTemplate.Utilities.Configuration
+namespace SeleniumSharpTemplate.Utilities.Configuration;
+
+public static class ConfigurationFactory
 {
-    public static class ConfigurationFactory
-    {
-        public static T? GetBinding<T>(string jsonFileName, string section) =>
-            new ConfigurationBuilder()
-            .AddJsonFile(jsonFileName)
-            .Build()
-            .GetSection(section)
-            .Get<T>();
-    }
+    public static T? GetBinding<T>(string jsonFileName, string section) =>
+        new ConfigurationBuilder()
+        .AddJsonFile(jsonFileName)
+        .Build()
+        .GetSection(section)
+        .Get<T>();
 }
