@@ -1,9 +1,8 @@
-﻿namespace SeleniumSharpTemplate.Utilities.Exceptions
+﻿namespace SeleniumSharpTemplate.Utilities.Exceptions;
+
+public static class Messages
 {
-    public static class Messages
-    {
-        public const string BROWSER_EX_MESSAGE = "Invalid browser type specified";
-        public const string DRIVER_EX_MESSAGE = "Invalid driver type specified";
-        public const string SCREENSHOT_EX_MESSAGE = "Unable to take screenshot";
-    }
+    public const string BROWSER_EX_MESSAGE = "Invalid browser type specified";
+    public const string DRIVER_EX_MESSAGE = "Invalid driver type specified";
+    public const string SCREENSHOT_EX_MESSAGE = "Unable to take screenshot";
 }
