@@ -1,13 +1,12 @@
-﻿namespace SeleniumSharpTemplate.Utilities.Paths
-{
-    public static class PathResolver
-    {
-        public static string CreatePathIfNotExists(this string path)
-        {
-            DirectoryInfo directoryInfo = new(path);
-            if (!directoryInfo.Exists) directoryInfo.Create();
+﻿namespace SeleniumSharpTemplate.Utilities.Paths;
 
-            return directoryInfo.FullName;
-        }
+public static class PathResolver
+{
+    public static string CreatePathIfNotExists(this string path)
+    {
+        DirectoryInfo directoryInfo = new(path);
+        if (!directoryInfo.Exists) directoryInfo.Create();
+
+        return directoryInfo.FullName;
     }
 }
