@@ -3,14 +3,13 @@ using SeleniumSharpTemplate.Utilities.WebDrivers.Abstract;
 using SeleniumSharpTemplate.Utilities.WebDrivers.Options.Chrome;
 using SeleniumSharpTemplate.Utilities.WebDrivers.ServiceCreator.Chrome;
 
-namespace SeleniumSharpTemplate.Utilities.WebDrivers.Initializer.Chrome
-{
-    public class ChromeWebDriverInitializer : WebDriverInitializerBase
-    {
-        public override IWebDriver InitializeDriver() =>
-            new ChromeDriver(ChromeDriverServiceCreator.Service, ChromeDriverOptions.Options);
+namespace SeleniumSharpTemplate.Utilities.WebDrivers.Initializer.Chrome;
 
-        public override IWebDriver InitializeHeadlessDriver() =>
-            new ChromeDriver(ChromeDriverServiceCreator.Service, ChromeDriverOptions.HeadlessOptions);
-    }
+public class ChromeWebDriverInitializer : WebDriverInitializerBase
+{
+    public override IWebDriver InitializeDriver() =>
+        new ChromeDriver(ChromeDriverServiceCreator.Service, ChromeDriverOptions.Options);
+
+    public override IWebDriver InitializeHeadlessDriver() =>
+        new ChromeDriver(ChromeDriverServiceCreator.Service, ChromeDriverOptions.HeadlessOptions);
 }
