@@ -1,26 +1,25 @@
 ﻿using OpenQA.Selenium.Firefox;
 
-namespace SeleniumSharpTemplate.Utilities.WebDrivers.ServiceCreator.Firefox
+namespace SeleniumSharpTemplate.Utilities.WebDrivers.ServiceCreator.Firefox;
+
+public static class FirefoxDriverServiceCreator
 {
-    public static class FirefoxDriverServiceCreator
+    private const string SE_CACHE_PATH = "SE_CACHE_PATH";
+
+    public static FirefoxDriverService Service
     {
-        private const string SE_CACHE_PATH = "SE_CACHE_PATH";
-
-        public static FirefoxDriverService Service
+        get
         {
-            get
-            {
-                DirectoryInfo driverPath =
-                    new($"{Path.Combine(PathFinder.Bin, FileAndFolderName.DRIVER)}");
-                if (!driverPath.Exists) driverPath.Create();
+            DirectoryInfo driverPath =
+                new($"{Path.Combine(PathFinder.Bin, FileAndFolderName.DRIVER)}");
+            if (!driverPath.Exists) driverPath.Create();
 
-                Environment.SetEnvironmentVariable(SE_CACHE_PATH, driverPath.ToString());
+            Environment.SetEnvironmentVariable(SE_CACHE_PATH, driverPath.ToString());
 
-                FirefoxDriverService service = FirefoxDriverService.CreateDefaultService();
-                service.Start();
+            FirefoxDriverService service = FirefoxDriverService.CreateDefaultService();
+            service.Start();
 
-                return service;
-            }
+            return service;
         }
     }
 }
